requirements.txt
packaging~=25.0
setuptools~=80.9.0
PyYAML~=6.0.2
tqdm~=4.67.1
gradio~=5.43.1
python-dotenv~=1.1.1
ollama~=0.5.3
langchain~=0.3.27
langchain-core~=0.3.74
langchain-community~=0.3.27
pdfplumber~=0.11.7
PyPDF2~=3.0.1
sentence-transformers~=5.1.0
tabula-py~=2.10.0